requirements.txt
aiohttp==3.11.14
discord.py==2.5.2
python-dotenv==1.1.1
requests==2.32.3